+++ src/components/sidebar/helpers/generate-top-level-sidebar-nav-data.ts
