+++ src/components/sidebar/helpers/index.ts
@@ -1,6 +1,6 @@
 import { addNavItemMetaData } from './add-nav-item-meta-data'
 import { generateProductLandingSidebarNavData } from './generate-product-landing-nav-items'
-import { generateTopLevelSidebarNavData } from './generate-top-level-nav-items'
+import { generateTopLevelSidebarNavData } from './generate-top-level-sidebar-nav-data'
 import { getFilteredNavItems } from './get-filtered-nav-items'
 
 export {
